+++ src/main/java/org/folio/bulkops/batch/MarcAsListStringsWriter.java
@@ -8,6 +8,7 @@ import lombok.extern.log4j.Log4j2;
 import org.folio.bulkops.client.SrsClient;
 import org.folio.bulkops.domain.bean.BulkOperationsEntity;
 import org.folio.bulkops.domain.bean.ExtendedInstance;
+import org.folio.bulkops.domain.converter.JsonToMarcConverter;
 import org.folio.bulkops.domain.dto.ErrorType;
 import org.folio.bulkops.exception.BulkEditException;
 import org.springframework.batch.core.configuration.annotation.StepScope;
