+++ src/main/java/org/folio/bulkops/batch/jobs/BulkEditHoldingsProcessor.java
@@ -26,6 +26,7 @@ import org.folio.bulkops.domain.dto.IdentifierType;
 import org.folio.bulkops.exception.BulkEditException;
 import org.folio.bulkops.processor.EntityExtractor;
 import org.folio.bulkops.processor.permissions.check.PermissionsValidator;
+import org.folio.bulkops.processor.permissions.check.TenantResolver;
 import org.folio.bulkops.service.HoldingsReferenceService;
 import org.folio.bulkops.batch.jobs.processidentifiers.DuplicationCheckerFactory;
 import org.folio.bulkops.client.HoldingsClient;
