+++ src/main/java/org/folio/bulkops/batch/jobs/processidentifiers/BulkEditInstanceIdentifiersJobConfig.java
@@ -17,7 +17,6 @@ import lombok.extern.log4j.Log4j2;
 import org.folio.bulkops.batch.BulkEditSkipListener;
 import org.folio.bulkops.batch.JobCompletionNotificationListener;
 import org.folio.bulkops.batch.JsonListFileWriter;
-import org.folio.bulkops.batch.JsonToMarcConverter;
 import org.folio.bulkops.batch.MarcAsListStringsWriter;
 import org.folio.bulkops.batch.CsvListItemWriter;
 import org.folio.bulkops.batch.jobs.BulkEditInstanceProcessor;
@@ -25,6 +24,7 @@ import org.folio.bulkops.client.RemoteFileSystemClient;
 import org.folio.bulkops.client.SrsClient;
 import org.folio.bulkops.domain.bean.ExtendedInstance;
 import org.folio.bulkops.domain.bean.ItemIdentifier;
+import org.folio.bulkops.domain.converter.JsonToMarcConverter;
 import org.folio.bulkops.domain.dto.EntityType;
 import org.folio.bulkops.exception.BulkEditException;
 import org.folio.bulkops.service.ErrorService;
